\name{eye}
\alias{eye}
\title{
    Identity Matrix
}
\description{
    Returns an n x n square matrix with ones on the main diagonal and zeros elsewhere.
}
\usage{
    eye(n)
}
\arguments{
  \item{n}{
    the size of both matrix dimensions
  }
}
\value{
    Returns an n x n square matrix with ones on the main diagonal and zeros elsewhere.
}
\author{
    Bill Cruise
}