\name{mtrx-package}
\alias{mtrx-package}
\alias{mtrx}
\docType{package}
\title{
    Matrix creation convenience functions.
}
\description{
    Matrix creation convenience functions. The function names and parameters will be familiar if you're coming from Octave.
}
\details{
\tabular{ll}{
Package: \tab mtrx\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-20\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Bill Cruise

Maintainer: Bill Cruise <cruise02@gmail.com>
}
