\name{randi}
\alias{randi}
\title{
    Random Integer Matrix
}
\description{
    Creates an m x n matrix with random integer elements.
}
\usage{
    randi(imax, m, n = m, seed = NULL)
}
\arguments{
  \item{imax}{
    maximum integer value in the range
  }
  \item{m}{
    number of rows in the matrix
  }
  \item{n}{
    number of columns in the matrix
  }
  \item{seed}{
    initial seed value for the random generator
  }
}
\value{
    Returns an m x n matrix with random integer elements.
}
\author{
    Bill Cruise
}
