\name{zeros}
\alias{zeros}
\title{
    Zeros Matrix
}
\description{
    Creates a matrix or N-dimensional array whose elements are all 0.
}
\usage{
    zeros(m, n = m)
}
\arguments{
  \item{m}{
    number of rows in the matrix
  }
  \item{n}{
    number of columns in the matrix
  }
}
\details{
    If invoked with a single integer argument, creates a square matrix.
}
\value{
    Returns a matrix or N-dimensional array whose elements are all 0.
}
\author{
    Bill Cruise
}
